scripts/requirements.txt
# Python dependencies for GA4 property discovery script

# Required for GA4 Admin API
google-analytics-admin>=0.22.0
google-auth>=2.23.0

# Optional for direct Supabase insertion
supabase>=2.0.0

# For better output formatting
colorama>=0.4.6
